fix playwright browsers cache dir
Set the PLAYWRIGHT_BROWSERS_PATH environment variable to point to the
correct cache directory (`/root/.cache/ms-playwright/`) in the GitLab CI
pipeline.

This ensures that Playwright uses the pre-installed browsers from the
cache preloaded in the deps-frontend image instead of redownloading them,
speeding up the CI process and reducing network usage.

This creates a /root/.cache/ms-playwright/ dir in the resulting image, with this content: